api/requirements.txt
# FastAPI Backend Requirements

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# ONNX Inference
onnxruntime>=1.15.0
numpy>=1.24.0

# Image Processing
Pillow>=9.5.0

# CORS and utilities
starlette>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client (for health checks)
httpx>=0.24.0

# Optional: For development
python-dotenv>=1.0.0

